Directly use the trainer name in the battle end text
The original ROM has some code to copy a shortened version of the
trainer name to a temporary buffer before printing the battle end text.
Apparently in the English version this is redundant because the
shortened versions of all the names are the same. Let’s remove all this
code so we don’t have to maintain two copies of all the trainer names.